fix httpRequest
