+++ lib/client/kadira.js
@@ -4,6 +4,7 @@ import { Meteor } from 'meteor/meteor';
 import { getErrorParameters } from '../common/utils';
 import { ErrorModel } from './models/errors';
 import { getBrowserInfo } from './utils';
+import { httpRequest } from './httpRequest';
 
 Kadira.enableErrorTracking = function () {
   Kadira.options.enableErrorTracking = true;
@@ -90,3 +91,5 @@ if (Meteor.isCordova) {
   // eslint-disable-next-line camelcase
   initialize(__meteor_runtime_config__.kadira);
 }
+
+Kadira._makeHttpRequest = httpRequest;
