+++ lib/common/send.js
@@ -43,8 +43,7 @@ Kadira._getSendFunction = function () {
 };
 
 Kadira._clientSend = function (endpoint, payload, callback) {
-  import { httpRequest } from '../client/utils';
-  httpRequest('POST', endpoint, {
+  Kadira._makeHttpRequest('POST', endpoint, {
     headers: {
       'Content-Type': 'application/json'
     },
