+++ lib/ntp.js
@@ -120,9 +120,7 @@ Ntp.prototype.getServerTime = function (callback) {
         callback(err);
       });
   } else {
-    import { httpRequest } from './client/utils';
-
-    httpRequest('GET', `${self.endpoint}?noCache=${new Date().getTime()}-${Math.random()}`, function (err, res) {
+    Kadira._makeHttpRequest('GET', `${self.endpoint}?noCache=${new Date().getTime()}-${Math.random()}`, function (err, res) {
       if (err) {
         callback(err);
       } else {
